requirements.txt
# Core geospatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
folium>=0.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Census API
requests>=2.31.0

# Snowflake
snowflake-connector-python>=3.5.0
snowflake-sqlalchemy>=1.5.0

# Environment management
python-dotenv>=1.0.0
